requirements.txt
Django
django-import-export
django-simpleui
Pillow
docutils
django-mdeditor
markdown
Pygments